8240363: Refactor Compile::Output() to its own Phase Reviewed-by: kvn, vlivanov
